requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
gunicorn>=21.2
whitenoise>=6.7
psycopg2-binary>=2.9 ; platform_system != "Windows"
